+++ Views/Admin/Index.cshtml
@@ -57,6 +57,7 @@
     </section>
 
     <script src="~/js/utility.js" asp-append-version="true"></script>
+    <script src="~/js/admin.js" asp-append-version="true"></script>
     <script src="~/lib/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
