+++ wwwroot/js/utility.js
@@ -77,21 +77,3 @@ function createImageInBase64ForEvents() {
         }
     });
 }
-
-function clearAllFieldsFromAdmin(conteudo) {
-    if (conteudo === "usuarios" || conteudo === "eventos" || conteudo === "noticias") {
-        $("#contentInput").val("");
-        $("#headlineInput").val("");
-        $("#titleInput").val("");
-        $("#newsImageBase64").val("");
-        $("#formFileSm").val("");
-
-        $("#nameInput").val("");
-        $("#descriptionInput").val("");
-        $("#addressInput").val("");
-        $("#dateInput").val("");
-        $("#timeInput").val("");
-        $("#formEventsFileSm").val("");
-        $("#eventsImageBase64").val("");
-    }
-}
